% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc.hifld.hosp-package.R
\docType{package}
\name{dc.hifld.hosp-package}
\alias{dc.hifld.hosp}
\alias{dc.hifld.hosp-package}
\title{dc.hifld.hosp: Social Data Commons: Hospitals Access Scores}
\description{
Allows user to easily get Hospital Access Scores data from the SDAD Data Commons. Data can be provided directly or in a file.
}
\author{
\strong{Maintainer}: Cesar Montalvo \email{cpm9w@virginia.edu}

Authors:
\itemize{
  \item Social and Decision Analytics Division (SDAD), Biocomplexity Institute, University of Virginia
}

}
\keyword{internal}
